refactor: extract model-visible output budget helpers
