+++ crates/ironclaw_host_runtime/src/first_party_tools/mod.rs
@@ -9,6 +9,7 @@ mod echo;
 mod http;
 mod json;
 mod memory;
+mod model_visible_output;
 mod schemas;
 mod shell;
 mod skill_management;
